Remove resolved dependencies from repository
Instead of keeping the dependencies checked in, it is better they are
resolved everytime someone wants to actually build the project. Once
the dependecies have stabilized in terms of changes and / or releases
we will really profit from the dependency mechanism that sencha gives
us.

For now this likely means that the dependencies of the project (GeoExt
and basepackage) need to be available for the sencha command at build
time.