Change replacement character for dynamic replacements in GNOS values
This changes the characters that mark a candidate for replacements with values obtained from
a gettabl object from {{ and }} to [[ and ]].

Additionally a small testsuite is added which will be filled on subsequent commits. To run
the test suite, issue `npm install` (in the src-folder) and then `npm test`. Alternatively
you can also open `src/test/index.html` in a browser.