Allow to add filterless layers via filter form
Previously when the filter window was created, we would show an informative
message that we couldn't display the filter form for layers without filters.

In order to allow users that don't know or care about the doublecllick
behaviour, we now show the filterform even if there isn't any filter defined in
the metadata. The filter form then only contains a button to actually add the
layer to the map.

See #1325 (https://redmine-koala.bfs.de/issues/1325)