Allow printouts without scalebar and northArrow
This build on top of https://github.com/terrestris/BasiGX/pull/74/files and
ensures that the setting of the northArrow and scalebar checkboxes have the
desired effect. To not show these elements, some colors are set to be
transparent and the northArrow image is set to a point to a transparent gif.

The URL to the gif-file can be set via appConfig.json (under the key
urls/print-transparent-img).

See https://redmine-koala.bfs.de/issues/1398