Support `xAxisMin` and `xAxisMax` for timeseries
This adds support for minimum and maximum dates available to set via the filters
of the timeseries charts.

For the `classic` toolkit this works as expected, but for the `modern` toolkit
this is only prepared and commented out. In the `modern` filters we have to use
`Ext.field.DatePicker` (`datepickerfield`), which neither has config options
`minValue`/`maxValue` nor appropriate getters and setters.

For the `classic` toolkit, the first layer used to setup the window is used to
setup the minimum and maximum values of the filter.